Move HeaderForward helpers to vmcp/headerforward
Lift BuildHeaderForwardTripper and its supporting types out of
pkg/vmcp/client/ into the new sibling package pkg/vmcp/headerforward/.
The helper now has two consumers — pkg/vmcp/client (capability-discovery
HTTP client) and pkg/vmcp/session/internal/backend (per-session client)
— and is not logically part of either. A pkg/vmcp/headerforward/wirefmt
subpackage already exists, so the namespace is established.

Pure mechanical relocation: no behavior change, no signature change,
existing tests still cover the helper. This makes future vMCP
anti-pattern #8 cleanup (consolidation of the two parallel HTTP-client
builders) more straightforward.

Related to #5289 (the bug whose fix in PR #5301 motivated the export).
